Create a new grid from the size of the user's input
I changed the setting of the grid item's size to be a function
that will recieve an itemsAmount argument in order to create
a greed based on the argument provided. I then wrote a new
function resizeGrid that will recieve a variable from the user
and pass it to setGrid().