Move settings section to the left of the grid
I had to come up with a way to center the grid while having
the settings section on it's left, so I added an empty div
to the right of the grid and set it's width to be the same
as the settings section. I then put the side margings of the
grid on auto, this way it will always be centered.